+++ example/lib/main.dart
@@ -2,10 +2,12 @@ import 'package:accordion/accordion.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
-  runApp(AccordionApp());
+  runApp(const AccordionApp());
 }
 
 class AccordionApp extends StatelessWidget {
+  const AccordionApp({Key key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -14,47 +16,60 @@ class AccordionApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: AccordionPage(),
+      home: const AccordionPage(),
     );
   }
 }
 
 class AccordionPage extends StatelessWidget //__
 {
-  final _headerStyle = TextStyle(
+  const AccordionPage({Key key}) : super(key: key);
+
+  final _headerStyle = const TextStyle(
       color: Color(0xffffffff), fontSize: 15, fontWeight: FontWeight.bold);
-  final _contentStyleHeader = TextStyle(
+  final _contentStyleHeader = const TextStyle(
       color: Color(0xff999999), fontSize: 14, fontWeight: FontWeight.w700);
-  final _contentStyle = TextStyle(
+  final _contentStyle = const TextStyle(
       color: Color(0xff999999), fontSize: 14, fontWeight: FontWeight.normal);
   final _loremIpsum =
       '''Lorem ipsum is typically a corrupted version of 'De finibus bonorum et malorum', a 1st century BC text by the Roman statesman and philosopher Cicero, with words altered, added, and removed to make it nonsensical and improper Latin.''';
 
+  @override
   build(context) => Scaffold(
         backgroundColor: Colors.blueGrey[100],
         appBar: AppBar(
-          title: Text('Accordion'),
+          title: const Text('Accordion'),
         ),
         body: Accordion(
-          maxOpenSections: 3,
-          headerPadding: EdgeInsets.symmetric(vertical: 7, horizontal: 15),
+          maxOpenSections: 1,
+          headerBackgroundColorOpened: Colors.black54,
+          headerPadding:
+              const EdgeInsets.symmetric(vertical: 7, horizontal: 15),
+          // sectionOpeningHapticFeedback: SectionHapticFeedback.heavy,
+          // sectionClosingHapticFeedback: SectionHapticFeedback.light,
+          // openAndCloseAnimation: false,
           children: [
             AccordionSection(
               isOpen: true,
-              leftIcon: Icon(Icons.insights_rounded, color: Colors.white),
+              leftIcon: const Icon(Icons.insights_rounded, color: Colors.white),
+              headerBackgroundColor: Colors.black,
+              headerBackgroundColorOpened: Colors.red,
               header: Text('Introduction', style: _headerStyle),
               content: Text(_loremIpsum, style: _contentStyle),
               contentHorizontalPadding: 20,
               contentBorderWidth: 1,
+              // sectionOpeningHapticFeedback: SectionHapticFeedback.heavy,
+              // sectionClosingHapticFeedback: SectionHapticFeedback.vibrate,
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.compare_rounded, color: Colors.white),
+              leftIcon: const Icon(Icons.compare_rounded, color: Colors.white),
               header: Text('About Us', style: _headerStyle),
-              contentBorderColor: Color(0xffffffff),
+              contentBorderColor: const Color(0xffffffff),
+              headerBackgroundColorOpened: Colors.amber,
               content: Row(
                 children: [
-                  Icon(Icons.compare_rounded,
+                  const Icon(Icons.compare_rounded,
                       size: 120, color: Colors.orangeAccent),
                   Flexible(
                       flex: 1, child: Text(_loremIpsum, style: _contentStyle)),
@@ -63,7 +78,7 @@ class AccordionPage extends StatelessWidget //__
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.food_bank, color: Colors.white),
+              leftIcon: const Icon(Icons.food_bank, color: Colors.white),
               header: Text('Company Info', style: _headerStyle),
               content: DataTable(
                 sortAscending: true,
@@ -123,37 +138,39 @@ class AccordionPage extends StatelessWidget //__
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.contact_page, color: Colors.white),
+              leftIcon: const Icon(Icons.contact_page, color: Colors.white),
               header: Text('Contact', style: _headerStyle),
               content: Wrap(
                 children: List.generate(
                     30,
-                    (index) => Icon(Icons.contact_page,
+                    (index) => const Icon(Icons.contact_page,
                         size: 30, color: Color(0xff999999))),
               ),
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.computer, color: Colors.white),
+              leftIcon: const Icon(Icons.computer, color: Colors.white),
               header: Text('Jobs', style: _headerStyle),
-              content:
-                  Icon(Icons.computer, size: 200, color: Color(0xff999999)),
+              content: const Icon(Icons.computer,
+                  size: 200, color: Color(0xff999999)),
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.movie, color: Colors.white),
+              leftIcon: const Icon(Icons.movie, color: Colors.white),
               header: Text('Culture', style: _headerStyle),
-              content: Icon(Icons.movie, size: 200, color: Color(0xff999999)),
+              content:
+                  const Icon(Icons.movie, size: 200, color: Color(0xff999999)),
             ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.people, color: Colors.white),
+              leftIcon: const Icon(Icons.people, color: Colors.white),
               header: Text('Community', style: _headerStyle),
-              content: Icon(Icons.people, size: 200, color: Color(0xff999999)),
+              content:
+                  const Icon(Icons.people, size: 200, color: Color(0xff999999)),
             ),
             // AccordionSection(
             //   isOpen: false,
-            //   leftIcon: Icon(Icons.person_add, color: Colors.white),
+            //   leftIcon: const Icon(Icons.person_add, color: Colors.white),
             //   header: Text('Accordion within Accordion', style: _headerStyle),
             //   contentHorizontalPadding: 15,
             //   contentVerticalPadding: 15,
@@ -192,9 +209,10 @@ class AccordionPage extends StatelessWidget //__
             // ),
             AccordionSection(
               isOpen: false,
-              leftIcon: Icon(Icons.map, color: Colors.white),
+              leftIcon: const Icon(Icons.map, color: Colors.white),
               header: Text('Map', style: _headerStyle),
-              content: Icon(Icons.map, size: 200, color: Color(0xff999999)),
+              content:
+                  const Icon(Icons.map, size: 200, color: Color(0xff999999)),
             ),
           ],
         ),