+++ example/test/widget_test.dart
@@ -11,7 +11,7 @@ import 'package:flutter_test/flutter_test.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(AccordionApp());
+    await tester.pumpWidget(const AccordionApp());
 
     // Verify that our counter starts at 0.
     // expect(find.text('0'), findsOneWidget);